added mt.Monotime() (time.Duration)

+++ md/md_amd64.go
@@ -2,4 +2,4 @@
 
 package md
 
-const PageSize = 4096
+const PageSize = uint(4096)
